Print snoopi_deep flamegraph with pretty function syntax
Make the functions more readable by printing with a pretty-printed
function signature instead of a type tuple.

Note though that this isn't ideal since the type tuple ends up as the
_typeof_ the function, not the function instance. And not all callable
objects are singletons, so you cannot safely call instance on it.

Instead, we need to adjust the julia source to produce better records.