Add a display_type param to flamegraph() to allow aggregating results
E.g. for PProf, which can aggregate function calls into a call-graph, it
can be useful to print the frames as methods or even functions, rather
than as method instances.

Also, i've added a stub for `slottypes`, which i'd like to make the
default printing format, which is even more precise, and would include
constants being specialized on.