Dramatically improve `flamegraph(timing)` performance (#158)
* Dramatically improve `flamegraph(timing)` performance

Move method instance's specTypes Type Tuple from a compile-time argument
to a runtime argument, to prevent compiling a specialization for every
_value_ in the data!

I hadn't meant to write it that way in the first place, I had just done
it absent-mindedly, because `specTypes` is a Type, so I hadn't thought
about moving it back into the value domain.

We used `@snoopi_deep` to find out the problem and fix this performance
problem! It's neat to see it profiling itself! :tada:

* Add error-handling for malformed Type Tuples that fail to print.

This apparently can happen sometimes, see:
https://github.com/JuliaLang/julia/issues/38195