Add tools for analysis of specialization (#169)
Here, the idea is to compare the inference time against the runtime
for each MethodInstance. The user is responsible for coming up with a
suitable runtime performance task and collecting Profile data.

If one loads PyPlot, then `specialization_plot` becomes available and
one can interact by clicking on individual dots to see the corresponding
Method.