Add `inference_triggers` (#159)
This adds utilities that make it easier to analyze the causes of inference triggers.
In contrast to the rest of the analysis, time doesn't seem to be an important concept
here: it's mostly irrelevant whether the current or next inference is a long job or a short
one, the key question is really: what's breaking up inference into successive independent
runs, and what might be done about it. This provides a number of utilities that make it
easier to identify and examine the caller, and to gather statistics about the frequency
of triggering inference by callsite.