Skip higher-order funcs for inference triggers
When identifying the caller of an inference-triggering callee,
sometimes it's nice to skip over higher order functions.
For instance in a `do`-block, you might prefer to attribute
any inference trigger in the anonymous function to the parent
scope that implemented the `do`-block rather than the
higher-order function used by the `do`-statement.