`ascend` integration and docs cleanup
This provides integration with Cthulhu's `ascend`, making it
significantly easier to analyze inference triggers.

The docstrings had also gotten out of date, and this updates them
while also allowing more compact docstrings by moving definitions to a
set of demos.

Finally, this adds an "orientation" docstring to SnoopCompile itself.
These can be helpful in orienting users to the API.