Fix attribution of "delayed" invalidations (#301)
"Delayed" invalidations occur during package loading
in response to some event that happened before
the package was loaded. Because the trigger is separated
in time from its consequences, we did not formerly
have a mechanism to clearly attribute these to a cause.
Julia 1.9 has recently added more logging to make this
more robust; this PR updates SnoopCompile to take advantage
of the new information.