Move `flatten()` to SnoopCompileCore.
At RelationalAI, we are turning on always-on `@snoopi_deep` profiling in
our production servers. For this, we want to add a dependency on
SnoopCompileCore, but not on SnoopCompile, since SnoopCompile adds many
dependencies for visualization utilities, such as FlameGraphs.jl.

We think that `flatten()` is a small enough and simple enough utility
function that it could easily live in SnoopCompileCore. Does that sound
reasonable to you as well?

Thanks! :)