Update `@snoop_inference` for Julia 1.12 (#419)
This leverages the new timing fields of CodeInstance to replace the old
Core.Compiler.Timings module. There is a bit of loss particularly in the
domain of constant-propagation, which is illustrated in
https://gist.github.com/timholy/9a64b27c1932bb414e69b8fe48284b5e. (Read
testcase.jl before looking at the results.) As a consequence there are a
small handful of failing tests locally, but the vast majority pass.